ike-init: Ignore non-matching KE payloads and INVALID_KE_PAYLOAD notifies
In case we send retransmits for an IKE_SA_INIT where we propose a DH
group the responder will reject we might later receive delayed responses
that either contain INVALID_KE_PAYLOAD notifies with the group we already
use or, if we retransmitted an IKE_SA_INIT with the requested group but
then had to restart again, a KE payload with a group different from the
one we proposed.  We ignore these messages now so we don't increase
the retry counter (which would cause an abort as would working with a KE
payload that doesn't match our own DH instance).

The main issue is, of course, that we reuse the same initiator SPI when
starting from scratch.  So an alternative approach would be to use a new
initiator SPI, but changing that on the existing IKE_SA object is tricky
as the local SPI is the key in the IKE_SA hashtable.  And creating a new
IKE_SA object could also have negative side-effects in regards to things
like make-before-break reauthentication that e.g. rely on special queued
tasks (the reestablish_* hooks might also be called even though the
previous SA never was established).